Update brave to 0.62.50
Summary:
Most notable for Linux users is the support for Widevine which makes it possible to watch for example Netflix, Amazon Prime Video and listen to Spotify. Widevine is not enabled by default and will thus need to be enabled from an icon that pops up on a site that requests to use Widevine for drm content.
Full changelog for this big release can be found [here](https://github.com/brave/brave-browser/releases/tag/v0.62.50).

Test Plan:
- Verified that it was possible to surf the web.
- Verfiied that ad-blocking worked as expected.
- Tested new Widevine support on `play.spotify.com`.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Differential Revision: https://dev.getsol.us/D5927